compliance-agent/requirements.txt
lxml==6.0.2
python-docx==1.2.0
